requirements.txt
transformers
vllm
fire
evaluate
datasets
jiwer
more_itertools
peft
torchaudio
distro
autoawq
huggingface-hub
google-generativeai
openai